+++ public/locale/en.json
@@ -1696,6 +1696,7 @@
   "my_profile": "My Profile",
   "my_schedules": "My Schedules",
   "name": "Name",
+  "name_and_slug_are_required": "Name and slug are required",
   "name_is_required": "Name is required",
   "name_of_hospital": "Name of Hospital",
   "name_of_shifting_approving_facility": "Name of shifting approving facility",
@@ -2596,8 +2597,10 @@
   "system": "System",
   "systolic": "Systolic",
   "tachycardia": "Tachycardia",
+  "tag_created_successfully": "Tag created successfully",
   "tag_name": "Tag Name",
   "tag_slug": "Tag Slug",
+  "tag_updated_successfully": "Tag updated successfully",
   "tags": "Tags",
   "taken": "Taken",
   "taper_titrate_dosage": "Taper & Titrate Dosage",
